requirements.txt
annotated-types==0.7.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.4.2
blosc2==2.0.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.2.1
coverage==7.5.3
cycler==0.12.1
Cython==3.0.10
DateTime==5.5
decorator==5.1.1
dpath==2.1.6
fonttools==4.53.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h5py==3.11.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
ipython==7.34.0
jedi==0.19.1
joblib==1.4.2
kiwisolver==1.4.5
linecheck==0.1.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
matplotlib-label-lines==0.7.0
microdf-python==0.3.0
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
nptyping==1.4.4
numexpr==2.10.0
numpy==1.24.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
patsy==0.5.6
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
policyengine-core==2.21.8
policyengine-us==0.780.0
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
pytest-dependency==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
statsmodels==0.14.2
synthimpute==0.1
tables==3.8.0
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
typish==1.9.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
yaml-changelog==0.3.0
zope.interface==6.4.post2
